+++ src/marketdata/resources/utilities/status.py
@@ -31,7 +31,6 @@ def status(
         url="status/",
         include_api_version=False,
         check_rate_limits=False,
-        populate_rate_limits=False,
     )
 
     return render(